src/mmw/requirements/base.txt
django<4
gunicorn==20.1.0
django-redis==5.1.0
hiredis==2.0.0
djangorestframework==3.13.1
django-filter==21.1
django-registration-redux==2.9
django-extensions==3.1.5
python-omgeo==6.1.0
rauth==0.7.3
djangorestframework-gis==0.17.0
drf_yasg==1.20.0
django-cors-headers==3.10.1
cryptography==36.0.1
pyOpenSSL==21.0.0
markdown==3.3.6
tr55==1.3.0
gwlf-e==3.1.0
requests[security]==2.26.0
rollbar==0.16.2
retry==0.9.2
python-dateutil==2.8.2
suds-community==1.0.0
django_celery_results==2.2.0
pandas==1.3.5
ulmo==0.8.8
hs_restclient==1.3.7
fiona==1.8.20
redis<4
numpy==1.20.3
dictdiffer==0.9.0
BMPxlsx==3.0.0
